feat(castling): implement public castling interface
Castling is a relatively big feature. It requires adding:
1. State to track castling
2. Updating the state as moves happen, so that castling is disallowed
   wherever a rook or king moves
3. Actually figuring out the castling move

Also, castling involves checking for threat a lot more than other moves,
since it can't be done if the king is currently in check, and it also
can't be done if any of the squares the king would move over are
threatened.

I'm splitting this work into multiple commits. This commit implements
the public facing interface of generating castling moves, but with all
the private implementation left todo!()

This also hasn't started implementing state updates