docs: add LICENSE
I'm putting this project back on Github publically because some folks
have shown interest in contributing. Part of doing this is setting up a
license. I'm a developer and I have no idea what I'm doing.

I went with the Apache license because
1. It's pretty permissive. I don't really care what folks do with this
   project, if it's of any use to them at all.
2. It doesn't require me to attribute contributors specifically. The MIT
   license requires that you copy the attribution everywhere. I'd feel
   weird sticking just my name onto the copyright holders list, and
   trying to collect everyone's names is just awkward.

So this project is licensed under the Apache license now.