docs: add CONTRIBUTING.md
I've been semi-enforcing some standards for contributions recently, but
it's been a confusing and annoying process for those that aren't me.
This documentation should hopefully help people stick to the same set of
rules and guidelines for there patches, and also be a place for me to
point at when I want people to make changes.