fix: raycasts not working with capturing for black
Initially, the architecture of the code didn't use Side::Black and
Side::White. It was CurrentlyMoving and MovingNext. When it was
refactored to black and white, I missed refactoring this function.

The old implementation made sense before, obviously you'd only check if
the piece is CurrentlyMoving (now renamed to White).