docs: todo for pawn to move en_passant impl out
While implementing castling, I realized that a decent pattern was taking
all the implementation for a particular piece and putting it in that
piece's module. Right now, this is muddled by en_passant, which is
partially implemented in the pawn module, and partially in the base
implementation for board. This means that changes to pawn also make deep
changes to the board.

We should ideally move this implementation out to the pawn module
itself.