fix(pawn): assuming White to be the moving side
The pawn module was written when instead of storing sides by the White
and Black, they were CurrentlyMoving and MovingNext. This meant that the
code could assume that the side would always be CurrentlyMoving.

I changed this a while ago, because that design meant that on every move
flip, we would have had to redo the every piece as one of the other
side. I figured the cost of doing that would be too high

I just missed this in the pawn code, because all our tests up until now
only tested moves for White.