test: add test for total possible moves
A good way to test a chess engine is to ensure that it creates the
correct number of moves at a given ply. These numbers are easily
available online.

In this case, the engine is really slow, and can't really generate more
than 4 ply without slowing down significantly. However, it's still a
valuable test.