refactor: make movegen generate moves, not boards
Movegen was generating entire boards for each move. This meant that for
any consumers of movegen, they have to have memory for tons of boards to
store game trees. This is pretty inefficient, not to mention kind of
useless. UCI engines expect moves, not entire board states, and search
algorithms want to maintain board state internally using hashmaps. There
is almost no consumer that would want the boards directly.

For now, we start exposing a generate_moved_boards that is identical to
the old functionality, and introduce a generate_moves that just returns
ChessMoves. This involved a ton of refactoring:

1. Had to make every move generator return moves instead of boards. For
now, internally they still try to execute the move, creating their own
duplicate of the board, but they just return the generated move.

In the future, as we move to an architecture that checks if a move is
possible and then executes it, we could just reuse the "check" part of
that system and not have to make tons of copies.
2. Had add a new type of chess move (en passant capture). This was
required because otherwise, code to execute en passant moves would have
had to be moved into `board`. This way, it stays in pawn, and we just
get an extra move type.

Overall, this commit both enables writing search systems, and enables us
to go off and implement UCI support.

References: https://todo.sr.ht/~lady_interstellar/knight_witch/3
References: https://todo.sr.ht/~lady_interstellar/knight_witch/6