refactor: make number of moves test use hashmaps
The number of moves test is relatively slow. It takes a long time to
generate the move tree and store it in memory.

This commit adds a hashmap to store the generated moves, instead of
storing all the generated boards in memory directly. This means:

1. For the last iteration of the movegen, we don't actually have to
apply the moves.
2. All our game state types are now hashable, which should mean we can
use hashtables for caching board evaluations etc. in search algorithms