feat(board): add chess result
This should let evaluation factor in whether or not something is good or
not. This should also let us interact with humans better, we'll be able
to say when someone is in checkmate or in stalemate.

The current implementation is very inefficient, and relies on calling
both check_king_threat and generate_moves, both notoriously bad in our
project. Hopefully these will be optimized further, or we can cache
them.