feat(evaluation): add game result evaluator
I'm trying to make evaluators composable. This should lead to easier
experimentation and maintainance of the code. It'd be really hard to
test the combined behavior of a giant evaluator, but they

This involved a ton of refactoring.

First of all, we needed a way for an evaluation to return a "certain"
result, because centipawns are just an estimate.

Secondly, this also meant adding comparisions for these game results. I
decided to implement this in two stages: one in a ChessResult inside the
board module, and the other for Evaluations as an enum, that calls the
result comparision inline.

Thirdly, I also needed to add some macros to reduce code repetition in
evaluator tests.

I'm guessing that we'll end up with tons of little
evaluators (generic piece position evaluators, king safety evaluators,
etc. etc. etc.). Hopefully with these new utillities and with the
abstractions that we have set up, it'll be easy to add and tune
evaluators, at least the simple ones.