refactor(test): Improved test performance
The perf test was storing all generated boards in memory. That meant
that we were waiting really long on all the boards beings stored and the
memory being allocated.

Moving this into just a simple recursive call meant a massive speedup.
This also means that we can run this test to an additional depth when
running in release mode and still have it execute in reasonable time for
the pre-push/apply hooks.