perf: improve king threat by ~20x
The previous implementation of king threat checks was basically a full
move generation step, which was incredibly expensive.

I reimplemented it to be much simpler: we now just check for everything
that could check the king, and see if there's an opponent's piece there.
That results in so much of a performance improvement that the move
generation perf test can be extended to an extra depth, and still run in
~1 second on my machine.