feat(evaluation): add composite evaluator
The whole idea with evaluation was that it was going to be composable.
One evaluator would build on top of another, and they'd all be run
together (potentially even in parallel someday) over a board to figure
out its value.

With CompositeEvaluator, this can be done. Implementing this involved
implementing Sum and Add for Evaluation, which also meant that the mod
file for evaluation was getting too big. I moved it to its own module.

The Evaluation enum module can't be called evaluation, because that's
what the base module is also called. For now, I've called it
evaluation_result.