feat: add null move
This should give a more pleasant interface to call the engine and flip
sides instead of calling the flip side method.

This is also useful for implementing search, down the line. In some
cases, search could reach a situation where no more moves are possible.
In a situation like that, we want to just return the current evaluation
and the best move as the "null move"