fix(minimax): depth being incorrect
Minimax was returning the wrong depth for certain results. It would
return 0 depth even when we still had to make a move to get a certain
result like a checkmate.

This meant that instead of reporting "Checkmate in 3 ply" like it should
it would report "Checkmate in 2 ply". This is most egregious when the
certain result happens after 1 move, because then it wrongly suggests
that the certain result is the current board state.

While doing this, I also found that minimax was calling for checked
make_move. It doesn't need to do this, since the moves it gets are
generated by the engine and so should be correct. I removed the check.